feat(front-end): refact dropdown component
