+++ front-end/src/components/organisms/Dropdown.tsx
@@ -4,9 +4,10 @@ import styled from 'styled-components';
 import { IconsTypes } from '../atoms/Icon/IconTypes';
 import { useState } from 'react';
 import BoxContainer from '../atoms/BoxContainer';
+import Options from '../molecules/Options';
 
 
-interface DropdownProps {
+interface DropdownInterface {
     label: string;
     icon: IconsTypes;
     options: any[];
@@ -33,56 +34,12 @@ export const LabelContainer = styled.div`
   justify-content: space-between;
 `;
 
-export const OptionsContainer = styled.div`
-  position: absolute;
-  left: 0;
-  top: 64px;
-  border-radius: 8px;
-  box-shadow: 0px 4px 5px rgba(0, 0, 0, 0.14);
-  background-color: ${({ theme }) => theme.alternative_primary};
-  color: ${({ theme }) => theme.white};
-  width: 100%;
-  max-height: 11rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  z-index: 100;
-  overflow-y: scroll;
-`;
-
-export const Option = styled.button`
-  border: none;
-  background: transparent;
-  width: 100%;
-  padding: 16px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  cursor: pointer;
-  color: ${({ theme }) => theme.white};
-  
-  :hover {
-    background-color: ${({ theme }) => theme.secondary};
-    color: ${({ theme }) => theme.primary};
-  }
-`;
-
-
-const Dropdown = ({ label, icon, options, onSelect }: DropdownProps) => {
+const Dropdown = ({ label, icon, options, onSelect }: DropdownInterface) => {
     const [open, setOpen] = useState(false);
 
     const handleSelect = (option: string) => {
         onSelect(option);
         setOpen(false);
-    }
-
-    const renderOptions = () => {
-        return options.map((option) => (
-            <Option onClick={() => handleSelect(option)}>
-                <Text cta size='16px'>{option}</Text>
-            </Option>
-        ))
     }
 
     return (
@@ -106,12 +63,8 @@ const Dropdown = ({ label, icon, options, onSelect }: DropdownProps) => {
                     rotate={open}
                 />
 
-                {open &&
-                    <OptionsContainer>
-                        {renderOptions()}
-                    </OptionsContainer>
-                }
             </Container >
+            {open && <Options options={options} onSelect={handleSelect} />}
         </BoxContainer>
     );
 };
